Clean up threading files.
